Moved all execution files to the directory exec/
